Changing over to unsigned numerical types since Valve responses use it. Tired of overflows. Improving some response parsing and handling.

+++ SteamWebAPI2/Models/DOTA2/GameItemResultContainer.cs
@@ -5,17 +5,17 @@ namespace SteamWebAPI2.Models.DOTA2
 {
     internal class GameItem
     {
-        public int Id { get; set; }
+        public uint Id { get; set; }
         public string Name { get; set; }
-        public int Cost { get; set; }
+        public uint Cost { get; set; }
 
         [JsonProperty(PropertyName = "secret_shop")]
-        public int SecretShop { get; set; }
+        public ushort SecretShop { get; set; }
 
         [JsonProperty(PropertyName = "side_shop")]
-        public int SideShop { get; set; }
+        public ushort SideShop { get; set; }
 
-        public int Recipe { get; set; }
+        public ushort Recipe { get; set; }
 
         [JsonProperty(PropertyName = "localized_name")]
         public string LocalizedName { get; set; }
